Add library dependency for mediawik/syntax-highlight
The `dev-REL1_27` version of `mediawiki/syntax-highlight` on
packages.bluespice.com declares a dependency to `symfony/process`.

Unfortunately this is not evaluated in the build of
`BlueSpice\mediawiki@REL1_27` as `Extension:SyntaxHighlight` is not loaded
by composer but is a git-submodule. The apprpriate `vendor` directory
which is also a git-submodule, but it get's overwritten by the build process.

Therefore we redeclare this dependency on the DistributionConnector to make
sure it is loaded properly.